pyproject.toml
[project]
name = "llm4matedu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chonkie>=1.4.0",
    "ipykernel>=7.0.1",
    "loguru>=0.7.3",
    "openai>=2.5.0",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]
